% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contributor_info.R
\name{get_contributor_info}
\alias{get_contributor_info}
\title{Get data on publication authors}
\usage{
get_contributor_info(base_data)
}
\arguments{
\item{base_data}{data frame}
}
\value{
data frame
}
\description{
Get data on publication authors
}
\examples{
base_data <- data.frame(cristin_result_id = 1094009, contributors.url = "https://api.cristin.no/v2/results/1094009/contributors")
get_contributor_info(base_data)
}
