% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cristin_results.R
\name{get_cristin_results}
\alias{get_cristin_results}
\title{Get Cristin results}
\usage{
get_cristin_results(unit = NULL, institution = NULL,
  contributor = NULL, project_code = NULL, user = NULL,
  category = NULL, funding_source = NULL, created_since = NULL,
  created_before = NULL, published_since = NULL,
  published_before = NULL, per_page = 999, page = 1,
  fields = "all", analyse = "base_data")
}
\arguments{
\item{unit}{string}

\item{institution}{string}

\item{contributor}{string}

\item{project_code}{string}

\item{user}{string}

\item{category}{string}

\item{funding_source}{string}

\item{created_since}{string}

\item{published_since}{string}

\item{published_before}{string}

\item{per_page}{string}

\item{page}{string}

\item{fields}{string}

\item{analyse}{string}

\item{created_befor}{string}
}
\value{
data frame
}
\description{
Get Cristin results
}
\examples{
get_cristin_results(contributor = "25062")
}
