% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cristin_results.R
\name{unpack}
\alias{unpack}
\title{Handle list-columns with NULL values stemming from the JSON parsing}
\usage{
unpack(x, base_data)
}
\arguments{
\item{x}{string}

\item{base_data}{data frame}
}
\value{
data frame
}
\description{
Handle list-columns with NULL values stemming from the JSON parsing
}
\examples{
unpack("projects", base_data)
}
