% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cristin_results.R
\name{get_cristin_results}
\alias{get_cristin_results}
\title{Get results from the Cristin API}
\usage{
get_cristin_results(..., page = 1, per_page = 999, fields = "all", NVI = FALSE)
}
\arguments{
\item{...}{Arguments to pass on to Cristin API. See 'API parameters' below
for more information.}

\item{page}{Page number for pagination purposes, set to 1 by default}

\item{per_page}{Number of items per page, set to 999 by default}

\item{fields}{'all' by default, if omitted will return a minimal set of
fields}

\item{NVI}{Logical determining whether to return only peer-reviewed
publications or all results, set to 'FALSE' by default}
}
\value{
A tibble with results from the API call, with one row per result.
}
\description{
\code{get_cristin_results()} takes in search parameters for the
Cristin API and returns the results as a tibble
}
\section{API parameters}{
The Cristin API recognises many search parameters. You
can find the latest specification at the Cristin API
\href{https://api.cristin.no/v2/}{documentation page}.

Here are examples of the formatting of these parameters:\tabular{lll}{
   parameter \tab type \tab description \cr
   \strong{doi} \tab string \tab DOI name (e.g. doi=10.1000/123456) \cr
}

\itemize{
\item \strong{title} The title of the result
\item \strong{contributor} Author's name or Cristin person id
\item \strong{issn} The issn of the result
\item \strong{unit} Id (e.g. unit=185.53.18.10)
\item \strong{institution} Id (one number, e.g.
institution=185), name or acronym of the institution the contributors
belong to
\item \strong{user} A person's username in Cristin together with the
institution id separated by ':' 'askeladd:185' means Cristin user
'askeladd' from the institution '185'
\item \strong{category} Result category
code, see
\href{https://api.cristin.no/v2/results/categories?lang=en}{Cristin
category codes} for specification
\item \strong{published_since} Results
published since and inclusive the given year, (yyyy), e.g: 2005
\item \strong{published_before} Results published before and inclusive the given
year, (yyyy), e.g: 2017
\item \strong{created_since} Results created since and
inclusive the given date, (yyyy-mm-dd), e.g: 2005-03-17 *
\strong{created_before} Results created before and inclusive the given
date, (yyyy-mm-dd), e.g: 2005-03-17 * \strong{modified_since} Results
modified since and inclusive the given date, (yyyy-mm-dd), e.g: 2005-03-17
\itemize{
\item \strong{modified_before} Results modified before and inclusive the given
date, (yyyy-mm-dd), e.g: 2005-03-17 * \strong{year_reported} The year a
result was reported * \strong{project_code} Project code is the internal
reference number used by funding source * \strong{funding_source} Funding
source code e.g: NFR * \strong{funding} Funding source code e.g: NFR, and
project_code together separated by ':' 'NFR:1234' means 'funding_source':
'NFR' with 'project_code': '1234' * \strong{lang} Two letter language code
that determines value of fields such as institution name, one of 'en'
(default), 'nb', or 'nn'.
}
}
}

\examples{
get_cristin_results(contributor = "25062")
}
