% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cristin_results.R
\name{get_cristin_results}
\alias{get_cristin_results}
\title{Get results from the Cristin API}
\usage{
get_cristin_results(..., page = 1, per_page = 999, fields = "all", NVI = FALSE)
}
\arguments{
\item{...}{Arguments to pass on to Cristin API. See 'API parameters' below
for more information.}

\item{page}{Page number for pagination purposes, set to 1 by default}

\item{per_page}{Number of items per page, set to 999 by default}

\item{fields}{'all' by default, if omitted will return a minimal set of
fields}

\item{NVI}{Logical determining whether to return only peer-reviewed
publications or all results, set to 'FALSE' by default}
}
\value{
A tibble with results from the API call, with one row per result.
}
\description{
\code{get_cristin_results()} takes in search parameters for the
Cristin API and returns the results as a tibble.
}
\section{API parameters The Cristin API recognises many search parameters.}{
You can find the latest specification at the Cristin API
\href{https://api.cristin.no/v2/}{documentation page}.

Here are examples of the formatting of these parameters:\tabular{ll}{
   parameter \tab description \cr
   \strong{doi} \tab DOI name (e.g. doi=10.1000/123456) \cr
   \strong{title} \tab The title of the result \cr
   \strong{contributor} \tab Author's name or Cristin person id \cr
   \strong{issn} \tab The issn of the result \cr
   \strong{unit} \tab Id (e.g. unit=185.53.18.10) \cr
   \strong{institution} \tab Id (one number, e.g. institution=185), name or acronym of the institution the contributors belong to \cr
   \strong{user} \tab A person's username in Cristin together with the institution id separated by ':' askeladd:185' means Cristin user 'askeladd' from the institution '185' \cr
   \strong{category} \tab Result category code, see \href{https://api.cristin.no/v2/results/categories?lang=en}{Cristin category codes} for specification \cr
   \strong{published_since} \tab Results published since and inclusive the given year, (yyyy), e.g: 2005 \cr
   \strong{published_before} \tab Results published before and inclusive the given year, (yyyy), e.g: 2017 \cr
   \strong{created_since} \tab Results created since and inclusive the given date, (yyyy-mm-dd), e.g: 2005-03-17 \cr
   \strong{created_before} \tab Results created before and inclusive the given date, (yyyy-mm-dd), e.g: 2005-03-17 \cr
   \strong{modified_since} \tab Results modified since and inclusive the given date, (yyyy-mm-dd), e.g: 2005-03-17 \cr
   \strong{modified_before} \tab Results modified before and inclusive the given date, (yyyy-mm-dd), e.g: 2005-03-17 \cr
   \strong{year_reported} \tab The year a result was reported \cr
   \strong{project_code} \tab Project code is the internal reference number used by funding source \cr
   \strong{funding_source} \tab Funding source code e.g: NFR \cr
   \strong{funding} \tab Funding source code e.g: NFR, and project_code together separated by ':' 'NFR:1234' means 'funding_source': NFR' with 'project_code': '1234' \cr
   \strong{lang} \tab Two letter language code that determines value of fields such as institution name, one of 'en' (default), 'nb', or 'nn'. \cr
}
}

\examples{
get_cristin_results(contributor = "25062")
}
